Update README.md
An interactive machine learning web application with Streamlit and Scikit-learn to explore different datasets and classifier. According to the chosen dataset and classifier the model will be created and accuracy for the same will be calculated with or without applying PCA (Principal Component analysis). This web app also contains several metrics for plotting of data.